requirements.txt
azure-core==1.4.0
azure-storage-blob==12.3.0
certifi==2018.8.24
cffi==1.14.0
chardet==3.0.4
cryptography==2.9
cycler==0.10.0
databricks-connect==6.2.0
idna==2.9
isodate==0.6.0
kiwisolver==1.1.0
matplotlib==3.0.3
msrest==0.6.13
numpy==1.18.4
oauthlib==3.1.0
Pillow==7.1.2
py4j==0.10.7
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.13.0
urllib3==1.25.8
wincertstore==0.2
